feat: add middleware more routes
