+++ server/src/middlewares/argon.middleware.ts
@@ -2,6 +2,52 @@ import argon2 from "argon2";
 import type { RequestHandler } from "express";
 import jwt from "jsonwebtoken";
 import type { JwtPayload } from "jsonwebtoken";
+import userRepository from "../modules/user/userRepository";
+
+export const login: RequestHandler = async (req, res, next) => {
+  try {
+    const user = await userRepository.readByEmailWithPassword(req.body.email);
+
+    if (user == null) {
+      res.sendStatus(422);
+      return;
+    }
+
+    const verified = await argon2.verify(user.password, req.body.password);
+
+    if (verified) {
+      const { password, ...userWithoutHashedPassword } = user;
+
+      const myPayload: JwtPayload = {
+        firstname: user.firstname,
+        lastname: user.lastname,
+      };
+
+      const token = await jwt.sign(
+        myPayload,
+        process.env.APP_SECRET as string,
+        {
+          expiresIn: "36h",
+        },
+      );
+
+      res
+        .cookie("auth_token", token, {
+          secure: false,
+          httpOnly: true,
+          maxAge: 3600000,
+        })
+        .status(200)
+        .json({
+          message: "Bienvenue sur le site !",
+        });
+    } else {
+      res.sendStatus(422);
+    }
+  } catch (err) {
+    next(err);
+  }
+};
 
 const hashingOptions = {
   type: argon2.argon2id,
