+++ server/src/router.ts
@@ -31,24 +31,13 @@ router.get("/api/ingredients/:id", ingredientActions.read);
 
 /* ************************************************************************* */
 
-import { hashPassword } from "./middlewares/argon.middleware";
-// Define user-related routes
-import {
-  checkEmail,
-  checkEmailAndStoreUserData,
-} from "./middlewares/checkEmail.middleware";
-import { login } from "./modules/auth/authActions";
+import { hashPassword, login } from "./middlewares/argon.middleware";
+import {checkEmail, checkEmailAndStoreUserData} from "./middlewares/checkEmail.middleware";
 import userActions from "./modules/user/userActions";
 import validateUser from "./validation/userValidation";
 
 router.post("/api/login", checkEmailAndStoreUserData, login);
-router.post(
-  "/api/register",
-  validateUser,
-  checkEmail,
-  hashPassword,
-  userActions.add,
-);
+router.post("/api/register",validateUser,checkEmail, hashPassword, userActions.add,);
 
 /* ************************************************************************* */
